+++ com.unity.ugui/Runtime/TMP/TMP_ResourcesManager.cs
@@ -10,31 +10,6 @@ namespace TMPro
     public static class TMP_ResourceManager
     {
         // ======================================================
-        // TEXT SETTINGS MANAGEMENT
-        // ======================================================
-
-        private static TMP_Settings s_TextSettings;
-
-        internal static TMP_Settings GetTextSettings()
-        {
-            if (s_TextSettings == null)
-            {
-                // Try loading the TMP Settings from a Resources folder in the user project.
-                s_TextSettings = Resources.Load<TMP_Settings>("TextSettings"); // ?? ScriptableObject.CreateInstance<TMP_Settings>();
-
-                #if UNITY_EDITOR
-                if (s_TextSettings == null)
-                {
-                    // Open TMP Resources Importer to enable the user to import the TMP Essential Resources and option TMP Examples & Extras
-                    TMP_PackageResourceImporterWindow.ShowPackageImporterWindow();
-                }
-                #endif
-            }
-
-            return s_TextSettings;
-        }
-
-        // ======================================================
         // FONT ASSET MANAGEMENT - Fields, Properties and Functions
         // ======================================================
 
@@ -68,7 +43,6 @@ namespace TMPro
         [RuntimeInitializeOnLoadMethod(RuntimeInitializeLoadType.AfterAssembliesLoaded)]
         static void ResetStaticsOnLoad()
         {
-            s_TextSettings = default;
             s_FontAssetReferences.Clear();
             s_FontAssetNameReferenceLookup.Clear();
             s_FontAssetFamilyNameAndStyleReferenceLookup.Clear();